refs #395 : update - refactor data component to support multiple storage lists
