+++ src/components/backend-ai-data-view.ts
@@ -38,8 +38,8 @@ import 'weightless/textfield';
 import '@material/mwc-icon-button';
 
 import '../plastics/lablup-shields/lablup-shields';
+import './backend-ai-storage-list';
 import {default as PainKiller} from './backend-ai-painkiller';
-import tus from '../lib/tus';
 
 import {BackendAiStyles} from "./backend-ai-console-styles";
 import {IronFlex, IronFlexAlignment, IronPositioning} from "../plastics/layout/iron-flex-layout-classes";
@@ -51,40 +51,15 @@ export default class BackendAIData extends BackendAIPage {
   @property({type: Boolean}) is_admin = false;
   @property({type: Boolean}) authenticated = false;
   @property({type: String}) deleteFolderId = '';
-  @property({type: Object}) explorer = Object();
-  @property({type: Array}) explorerFiles = [];
-  @property({type: Array}) invitees = [];
-  @property({type: String}) selectedFolder = '';
-  @property({type: Array}) uploadFiles = [];
   @property({type: String}) vhost = '';
   @property({type: Array}) vhosts = [];
   @property({type: Array}) allowedGroups = [];
-  @property({type: Object}) fileListGrid = Object();
-  @property({type: Object}) notification = Object();
-  @property({type: Object}) deleteFileDialog = Object();
-  @property({type: Object}) indicator = Object();
   @property({type: Array}) allowed_folder_type = [];
-  @property({type: Boolean}) uploadFilesExist = false;
-  @property({type: Object}) _boundIndexRenderer = Object();
-  @property({type: Object}) _boundTypeRenderer = Object();
-  @property({type: Object}) _boundControlFolderListRenderer = Object();
-  @property({type: Object}) _boundControlFileListRenderer = Object();
-  @property({type: Object}) _boundPermissionViewRenderer = Object();
-  @property({type: Object}) _boundFileNameRenderer = Object();
-  @property({type: Object}) _boundCreatedTimeRenderer = Object();
-  @property({type: Object}) _boundPermissionRenderer = Object();
-  @property({type: Boolean}) _uploadFlag = true;
+  @property({type: Object}) notification = Object();
+  @property({type: Object}) indicator = Object();
 
   constructor() {
     super();
-    this._boundIndexRenderer = this.indexRenderer.bind(this);
-    this._boundTypeRenderer = this.typeRenderer.bind(this);
-    this._boundControlFolderListRenderer = this.controlFolderListRenderer.bind(this);
-    this._boundControlFileListRenderer = this.controlFileListRenderer.bind(this);
-    this._boundPermissionViewRenderer = this.permissionViewRenderer.bind(this);
-    this._boundFileNameRenderer = this.fileNameRenderer.bind(this);
-    this._boundCreatedTimeRenderer = this.createdTimeRenderer.bind(this);
-    this._boundPermissionRenderer = this.permissionRenderer.bind(this);
   }
 
   static get styles() {
@@ -317,18 +292,6 @@ export default class BackendAIData extends BackendAIPage {
       `];
   }
 
-  _toggleCheckbox() {
-    let buttons = this.shadowRoot.querySelectorAll(".multiple-action-buttons");
-    if (this.fileListGrid.selectedItems.length > 0) {
-      [].forEach.call(buttons, (e: HTMLElement) => {
-        e.style.display = 'block';
-      });
-    } else {
-      [].forEach.call(buttons, (e: HTMLElement) => {
-        e.style.display = 'none';
-      });
-    }
-  }
 
   render() {
     // language=HTML
@@ -347,36 +310,9 @@ export default class BackendAIData extends BackendAIPage {
             New folder
           </wl-button>
         </h3>
-
-        <vaadin-grid class="folderlist" theme="row-stripes column-borders compact" aria-label="Folder list" .items="${this.folders}">
-          <vaadin-grid-column width="40px" flex-grow="0" resizable header="#" .renderer="${this._boundIndexRenderer}">
-          </vaadin-grid-column>
-          <vaadin-grid-column resizable header="Name">
-            <template>
-              <div class="indicator" @click="[[_folderExplorer()]]" .folder-id="[[item.name]]">[[item.name]]</div>
-            </template>
-          </vaadin-grid-column>
-
-          <vaadin-grid-column resizable>
-            <template class="header">id</template>
-            <template>
-              <div class="layout vertical">
-                <span class="indicator monospace">[[item.id]]</span>
-              </div>
-            </template>
-          </vaadin-grid-column>
-
-          <vaadin-grid-column width="85px" flex-grow="0" resizable header="Location">
-            <template>
-              <div class="layout vertical">
-                <span>[[item.host]]</span>
-              </div>
-            </template>
-          </vaadin-grid-column>
-          <vaadin-grid-column width="45px" flex-grow="0" resizable header="Type" .renderer="${this._boundTypeRenderer}"></vaadin-grid-column>
-          <vaadin-grid-column width="85px" flex-grow="0" resizable header="Permission" .renderer="${this._boundPermissionViewRenderer}"></vaadin-grid-column>
-          <vaadin-grid-column resizable header="Control" .renderer="${this._boundControlFolderListRenderer}"></vaadin-grid-column>
-        </vaadin-grid>
+        <div id="folder-lists" class="tab-content">
+          <backend-ai-storage-list active></backend-ai-storage-list>
+        </div>
       </wl-card>
       <wl-dialog id="add-folder-dialog" class="dialog-ask" fixed backdrop blockscrolling>
         <wl-card elevation="1" class="login-panel intro centered">
@@ -428,338 +364,7 @@ export default class BackendAIData extends BackendAIPage {
           </section>
         </wl-card>
       </wl-dialog>
-      <wl-dialog id="delete-folder-dialog" class="dialog-ask" fixed backdrop blockscrolling>
-        <wl-card class="login-panel intro centered">
-          <h3 class="horizontal center layout">
-            <span>Delete a folder</span>
-            <div class="flex"></div>
-            <wl-button fab flat inverted @click="${(e) => this._hideDialog(e)}">
-              <wl-icon>close</wl-icon>
-            </wl-button>
-          </h3>
-          <section>
-            <div class="warning">WARNING: this cannot be undone!</div>
-            <div>
-              <mwc-textfield class="red" id="delete-folder-name" label="Type folder name to delete"
-                           pattern="[a-zA-Z0-9_-]*"
-                           error-message="Allows letters, numbers and -_." auto-validate></mwc-textfield>
-              <br/>
-              <wl-button class="blue button" type="submit" id="delete-button" outlined @click="${() => this._deleteFolderWithCheck()}">
-                <wl-icon>close</wl-icon>
-                Delete
-              </wl-button>
-            </div>
-            </section>
-        </wl-card>
-      </wl-dialog>
-      <wl-dialog id="info-folder-dialog" class="dialog-ask" fixed backdrop blockscrolling>
-        <wl-card class="intro centered" style="margin: 0;">
-          <h3 class="horizontal center layout" style="border-bottom:1px solid #ddd;">
-            <span>${this.folderInfo.name}</span>
-            <div class="flex"></div>
-            <wl-button fab flat inverted @click="${(e) => this._hideDialog(e)}">
-              <wl-icon>close</wl-icon>
-            </wl-button>
-          </h3>
-          <div role="listbox" style="margin: 0;">
-            <vaadin-item>
-              <div><strong>ID</strong></div>
-              <div secondary>${this.folderInfo.id}</div>
-            </vaadin-item>
-            <vaadin-item>
-              <div><strong>Location</strong></div>
-              <div secondary>${this.folderInfo.host}</div>
-            </vaadin-item>
-            <vaadin-item>
-              <div><strong>Number of Files</strong></div>
-              <div secondary>${this.folderInfo.numFiles}</div>
-            </vaadin-item>
-            ${this.folderInfo.is_owner ? html`
-              <vaadin-item>
-                <div><strong>Ownership</strong></div>
-                <div secondary>You are the owner of this folder.</div>
-              </vaadin-item>
-            ` : html``}
-            <vaadin-item>
-              <div><strong>Permission</strong></div>
-              <div secondary>${this.folderInfo.permission}</div>
-            </vaadin-item>
-          </div>
-        </wl-card>
-      </wl-dialog>
-      <wl-dialog id="folder-explorer-dialog" class="folder-explorer">
-        <wl-card>
-          <h3 class="horizontal center layout" style="font-weight:bold">
-            <span>${this.explorer.id}</span>
-            <div class="flex"></div>
-            <wl-button fab flat inverted @click="${(e) => this._hideDialog(e)}">
-              <wl-icon>close</wl-icon>
-            </wl-button>
-          </h3>
-
-          <div class="breadcrumb">
-          ${this.explorer.breadcrumb ? html`
-              ${this.explorer.breadcrumb.map(item => html`
-               <wl-icon>keyboard_arrow_right</wl-icon>
-               <wl-button outlined class="goto" path="item" @click="${(e) => this._gotoFolder(e)}" dest="${item}">${item}</wl-button>
-              `)}
-              ` : html``}
-          </div>
-          <div class="horizontal layout folder-action-buttons">
-            <wl-button outlined class="multiple-action-buttons" @click="${() => this._openDeleteMultipleFileDialog()}" style="display:none;">
-              <div class="horizontal center layout">
-              <wl-icon style="--icon-size: 20px;margin-right:5px;">delete</wl-icon><span>Delete...</span></div>
-            </wl-button>
-            <wl-button outlined id="add-btn" @click="${(e) => this._uploadFileBtnClick(e)}">
-              <wl-icon style="--icon-size: 20px;margin-right:5px;">cloud_upload</wl-icon>
-              Upload Files...
-            </wl-button>
-            <wl-button outlined id="mkdir" @click="${() => this._mkdirDialog()}">
-              <wl-icon style="--icon-size: 20px;margin-right:5px;">create_new_folder</wl-icon>
-              New Folder
-            </wl-button>
-          </div>
-          <div id="dropzone"><p>drag</p></div>
-          <input type="file" id="fileInput" @change="${(e) => this._uploadFileChange(e)}" hidden multiple>
-          ${this.uploadFilesExist ? html`
-          <wl-button outlined id="cancel_upload" @click="${(e) => this._cancelUpload(e)}">
-            <wl-icon>cancel</wl-icon> Stop uploading
-          </wl-button>
-          <vaadin-grid class="progress" theme="row-stripes compact" aria-label="uploadFiles" .items="${this.uploadFiles}"
-                       height-by-rows>
-            <vaadin-grid-column width="100px" flex-grow="0">
-              <template>
-                <vaadin-item class="progress-item">
-                  <div>
-                    <template is="dom-if" if="[[item.complete]]">
-                      <wl-icon>check</wl-icon>
-                    </template>
-                  </div>
-                </vaadin-item>
-              </template>
-            </vaadin-grid-column>
-
-            <vaadin-grid-column>
-              <template>
-                <vaadin-item>
-                  <span>[[item.name]]</span>
-                  <template is="dom-if" if="[[!item.complete]]">
-                    <div>
-                      <vaadin-progress-bar value="[[item.progress]]"></vaadin-progress-bar>
-                    </div>
-                    <div>
-                      <span>[[item.caption]]</span>
-                    </div>
-                  </template>
-                </vaadin-item>
-              </template>
-            </vaadin-grid-column>
-          </vaadin-grid>` : html``}
-          <vaadin-grid id="fileList-grid" class="explorer" theme="row-stripes compact" aria-label="Explorer" .items="${this.explorerFiles}">
-            <vaadin-grid-selection-column auto-select></vaadin-grid-selection-column>
-            <vaadin-grid-column width="40px" flex-grow="0" resizable header="#" .renderer="${this._boundIndexRenderer}">
-            </vaadin-grid-column>
-
-            <vaadin-grid-sort-column flex-grow="2" resizable header="Name" path="filename" .renderer="${this._boundFileNameRenderer}">
-            </vaadin-grid-sort-column>
-
-            <vaadin-grid-sort-column flex-grow="2" resizable header="Created" path="ctime" .renderer="${this._boundCreatedTimeRenderer}">
-            </vaadin-grid-sort-column>
-
-            <vaadin-grid-column flex-grow="1" resizable>
-              <template class="header">
-                <vaadin-grid-sorter path="size">Size</vaadin-grid-sorter>
-              </template>
-              <template>
-                <div class="layout vertical">
-                  <span>[[item.size]]</span>
-                </div>
-              </template>
-            </vaadin-grid-column>
-            <vaadin-grid-column resizable flex-grow="2" header="Actions" .renderer="${this._boundControlFileListRenderer}"></vaadin-grid-column>
-          </vaadin-grid>
-        </wl-card>
-      </wl-dialog>
-
-      <wl-dialog id="mkdir-dialog" class="dialog-ask" fixed blockscrolling backdrop>
-        <wl-card elevation="1" class="login-panel intro centered" style="margin: 0;">
-          <h3 class="horizontal center layout">
-            <span>Create a new folder</span>
-            <div class="flex"></div>
-            <wl-button fab flat inverted @click="${(e) => this._hideDialog(e)}">
-              <wl-icon>close</wl-icon>
-            </wl-button>
-          </h3>
-          <section>
-            <mwc-textfield id="mkdir-name" label="Folder name" pattern="[a-zA-Z0-9_-]*" auto-validate
-                         error-message="Allows letters, numbers and -_."></mwc-textfield>
-            <br/>
-            <wl-button class="blue button" type="submit" id="mkdir-btn" @click="${(e) => this._mkdir(e)}" outlined>
-              <wl-icon>rowing</wl-icon>
-              Create
-            </wl-button>
-          </section>
-        </wl-card>
-      </wl-dialog>
-
-      <wl-dialog
-        id="share-folder-dialog"
-        class="dialog-ask"
-        fixed
-        backdrop
-        blockscrolling
-      >
-        <wl-card class="intro centered" style="margin: 0;">
-          <h3 class="horizontal center layout" style="border-bottom:1px solid #ddd;">
-            <span>Share Folder</span>
-            <div class="flex"></div>
-            <wl-button fab flat inverted @click="${(e) => this._hideDialog(e)}">
-              <wl-icon>close</wl-icon>
-            </wl-button>
-          </h3>
-          <div role="listbox" style="margin: 0; padding: 20px 25px 25px 25px;">
-            <div style="margin: 10px 0px">People</div>
-            <div style="display: flex;">
-              <div id="textfields" style="flex-grow: 2">
-                <wl-textfield type="email" label="Enter e-mail address"></wl-textfield>
-              </div>
-              <div>
-                <wl-button fab flat @click="${(e) => this._addTextField(e)}">
-                  <wl-icon>add</wl-icon>
-                </wl-button>
-                <wl-button fab flat @click="${(e) => this._removeTextField(e)}">
-                  <wl-icon>remove</wl-icon>
-                </wl-button>
-              </div>
-            </div>
-            <div style="margin: 10px 0px">Permissions</div>
-            <div style="display: flex; justify-content: space-evenly;">
-              <wl-label>
-                <wl-checkbox checked disabled></wl-checkbox>
-                View
-              </wl-label>
-              <wl-label>
-                <wl-checkbox id="share-folder-write"></wl-checkbox>
-                Edit
-              </wl-label>
-            </div>
-
-            <wl-button
-              type="button"
-              outlined
-              id="share-button"
-              style="width: 100%; box-sizing: border-box;"
-              @click=${e => this._shareFolder(e)}
-            >
-              <wl-icon>share</wl-icon>
-              Share
-            </wl-button>
-          </div>
-        </wl-card>
-      </wl-dialog>
-      <wl-dialog
-        id="modify-permission-dialog"
-        class="dialog-ask"
-        fixed backdrop blockscrolling
-      >
-        <wl-card class="intro" style="margin: 0; width: 100%;">
-          <h3 class="horizontal center layout" style="border-bottom:1px solid #ddd;">
-            <span>Modify Permissions</span>
-            <div class="flex"></div>
-            <wl-button fab flat inverted @click="${(e) => this._hideDialog(e)}">
-              <wl-icon>close</wl-icon>
-            </wl-button>
-          </h3>
-          <div role="listbox" style="margin: 0; padding: 10px;">
-            <vaadin-grid theme="row-stripes column-borders compact" .items="${this.invitees}">
-              <vaadin-grid-column
-                width="30px"
-                flex-grow="0"
-                header="#"
-                .renderer="${this._boundIndexRenderer}"
-              ></vaadin-grid-column>
-              <vaadin-grid-column header="Invitee Email">
-                <template>
-                  <div>[[item.shared_to.email]]</div>
-                </template>
-              </vaadin-grid-column>
-              <vaadin-grid-column header="Permission" .renderer="${this._boundPermissionRenderer}">
-              </vaadin-grid-column>
-            </vaadin-grid>
-          </div>
-        </wl-card>
-        <div slot="footer">
-          <wl-button
-            type="button"
-            outlined
-            style="width: 100%; box-sizing: border-box;"
-            @click=${this._modifySharedFolderPermissions}
-          >
-            <wl-icon>check</wl-icon>
-            Save Changes
-          </wl-button>
-        </div>
-      </wl-dialog>
-      <wl-dialog id="delete-file-dialog" fixed backdrop blockscrolling>
-         <wl-title level="3" slot="header">Let's double-check</wl-title>
-         <div slot="content">
-            <p>This action cannot be undone. Do you want to proceed?</p>
-         </div>
-         <div slot="footer">
-            <wl-button inverted flat @click="${(e) => this._hideDialog(e)}">Cancel</wl-button>
-            <wl-button @click="${(e) => this._deleteFileWithCheck(e)}">Okay</wl-button>
-         </div>
-      </wl-dialog>
-
     `;
-  }
-
-  _modifySharedFolderPermissions() {
-    const selectNodeList = this.shadowRoot.querySelectorAll('#modify-permission-dialog wl-select');
-    const inputList = Array.prototype.filter.call(selectNodeList, (pulldown, idx) => pulldown.value !== (this.invitees as any)[idx].perm)
-      .map((pulldown, idx) => ({
-        'perm': pulldown.value,
-        'user': (this.invitees as any)[idx].shared_to.uuid,
-        'vfolder': (this.invitees as any)[idx].vfolder_id
-      }));
-    const promiseArray = inputList.map(input => window.backendaiclient.vfolder.modify_invitee_permission(input));
-    Promise.all(promiseArray).then((response: any) => {
-      if (response.length === 0) {
-        this.notification.text = 'No changes made.';
-      } else {
-        this.notification.text = 'Permission successfully modified.';
-      }
-      this.notification.show();
-      this.shadowRoot.querySelector('#modify-permission-dialog').hide();
-    })
-  }
-
-  permissionRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div>
-          <wl-select outlined label="Select Permission">
-            <option ?selected=${rowData.item.perm === 'ro'} value="ro">View</option>
-            <option ?selected=${rowData.item.perm === 'rw'} value="rw">Edit</option>
-            <option ?selected=${rowData.item.perm === 'wd'} value="wd">Edit+Delete</option>
-          </wl-select>
-        </div>
-      `, root
-    )
-  }
-
-  _addTextField(e) {
-    let newTextField = document.createElement('wl-textfield');
-    newTextField.label = "Enter e-mail address";
-    newTextField.type = "email";
-
-    this.shadowRoot.querySelector('#textfields').appendChild(newTextField)
-  }
-
-  _removeTextField(e) {
-    const textfields = this.shadowRoot.querySelector('#textfields');
-    textfields.removeChild(textfields.lastChild);
   }
 
   indexRenderer(root, column?, rowData?) {
@@ -769,189 +374,9 @@ export default class BackendAIData extends BackendAIPage {
     );
   }
 
-  controlFolderListRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div
-          id="controls"
-          class="layout horizontal flex center"
-          folder-id="${rowData.item.name}"
-        >
-          <mwc-icon-button
-            class="fg green controls-running"
-            icon="info"
-            @click="${(e) => this._infoFolder(e)}"
-          ></mwc-icon-button>
-
-          ${this._hasPermission(rowData.item, 'r')
-        ? html`
-              <mwc-icon-button
-                class="fg blue controls-running"
-                icon="folder_open"
-                @click="${(e) => this._folderExplorer(e)}" .folder-id="${rowData.item.name}"
-              ></mwc-icon-button>
-            `
-        : html``
-      }
-
-          ${this._hasPermission(rowData.item, 'w') ? html`` : html``}
-
-          ${rowData.item.is_owner && rowData.item.type == 'user'
-        ? html`
-              <mwc-icon-button
-                class="fg blue controls-running"
-                icon="share"
-                @click="${(e) => this._shareFolderDialog(e)}"
-              ></mwc-icon-button>
-            `
-        : html``
-      }
-
-          ${rowData.item.is_owner
-        ? html`
-              <mwc-icon-button
-                class="fg cyan controls-running"
-                icon="perm_identity"
-                @click=${e => this._modifyPermissionDialog(rowData.item.id)}
-              ></mwc-icon-button>
-            `
-        : html``
-      }
-
-          ${this._hasPermission(rowData.item, 'd')
-        ? html`
-              <mwc-icon-button
-                class="fg red controls-running"
-                icon="delete"
-                @click="${(e) => this._deleteFolderDialog(e)}"
-              ></mwc-icon-button>
-            `
-        : html``
-      }
-        </div>
-       `, root
-    );
-  }
-
-  controlFileListRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        ${!this._isDir(rowData.item) && this._isDownloadable(rowData.item) ?
-        html`
-            <mwc-icon-button id="download-btn" class="tiny fg blue" icon="cloud_download"
-                               filename="${rowData.item.filename}" @click="${(e) => this._downloadFile(e)}"></mwc-icon-button>
-            <mwc-icon-button id="delete-btn" class="tiny fg red" icon="delete_forever"
-                               filename="${rowData.item.filename}" @click="${(e) => this._openDeleteFileDialog(e)}"></mwc-icon-button>
-                               ` : html``}
-       `, root
-    );
-  }
-
-  fileNameRenderer(root, column?, rowData?) {
-    render(
-      html`
-        ${this._isDir(rowData.item) ?
-        html`
-          <div class="indicator horizontal center layout" @click="${(e) => this._enqueueFolder(e)}" name="${rowData.item.filename}">
-            <mwc-icon-button class="fg controls-running" icon="folder_open"
-                               name="${rowData.item.filename}"></mwc-icon-button>
-            ${rowData.item.filename}
-          </div>
-       ` : html`
-          <div class="indicator horizontal center layout">
-            <mwc-icon-button class="fg controls-running" icon="insert_drive_file"></mwc-icon-button>
-            ${rowData.item.filename}
-          </div>
-       `}
-      `, root
-    );
-  }
-
-  permissionViewRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div class="horizontal center-justified wrap layout">
-        ${this._hasPermission(rowData.item, 'r') ? html`
-            <lablup-shields app="" color="green"
-                            description="R" ui="flat"></lablup-shields>` : html``}
-        ${this._hasPermission(rowData.item, 'w') ? html`
-            <lablup-shields app="" color="blue"
-                            description="W" ui="flat"></lablup-shields>` : html``}
-        ${this._hasPermission(rowData.item, 'd') ? html`
-            <lablup-shields app="" color="red"
-                            description="D" ui="flat"></lablup-shields>` : html``}
-        </div>`, root
-    )
-  }
-
-  createdTimeRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div class="layout vertical">
-            <span>${this._humanReadableTime(rowData.item.ctime)}</span>
-        </div>`, root
-    )
-  }
-
-  typeRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div class="layout vertical center-justified">
-        ${rowData.item.type == 'user' ? html`
-          <wl-icon>person</wl-icon>
-        ` : html`
-          <wl-icon>group</wl-icon>
-        `}
-        </div>`, root
-    )
-  }
-
   firstUpdated() {
-    this._addEventListenerDropZone();
-    this._mkdir = this._mkdir.bind(this);
-
-    this.deleteFileDialog = this.shadowRoot.querySelector('#delete-file-dialog');
-    this.fileListGrid = this.shadowRoot.querySelector('#fileList-grid');
-    this.fileListGrid.addEventListener('selected-items-changed', () => {
-      this._toggleCheckbox();
-    });
     this.indicator = this.shadowRoot.querySelector('#loading-indicator');
     this.notification = window.lablupNotification;
-    let textfields = this.shadowRoot.querySelectorAll('mwc-textfield');
-    for (const textfield of textfields) {
-      this._addInputValidator(textfield);
-    }
-    document.addEventListener('backend-ai-group-changed', (e) => this._refreshFolderList());
-    document.addEventListener('backend-ai-ui-changed', (e) => this._refreshFolderUI(e));
-    this._refreshFolderUI({"detail": {"mini-ui": window.mini_ui}});
-  }
-
-  _refreshFolderList() {
-    this.indicator.show();
-    let groupId = null;
-    groupId = window.backendaiclient.current_group_id();
-    let l = window.backendaiclient.vfolder.list(groupId);
-    l.then((value) => {
-      this.indicator.hide();
-      this.folders = value;
-    });
-    let vhosts = window.backendaiclient.vfolder.list_hosts();
-    vhosts.then((response) => {
-    });
-  }
-
-  _refreshFolderUI(e) {
-    let folder_explorer = this.shadowRoot.querySelector('#folder-explorer-dialog');
-    if (e.detail.hasOwnProperty('mini-ui') && e.detail['mini-ui'] === true) {
-      folder_explorer.classList.add('mini_ui');
-    } else {
-      folder_explorer.classList.remove('mini_ui');
-    }
   }
 
   async _viewStateChanged(active) {
@@ -966,7 +391,6 @@ export default class BackendAIData extends BackendAIPage {
         window.backendaiclient.vfolder.allowed_types().then(response => {
           this.allowed_folder_type = response;
         });
-        this._refreshFolderList();
       }, true);
     } else {
       this.is_admin = window.backendaiclient.is_admin;
@@ -974,7 +398,6 @@ export default class BackendAIData extends BackendAIPage {
       window.backendaiclient.vfolder.allowed_types().then(response => {
         this.allowed_folder_type = response;
       });
-      this._refreshFolderList();
     }
   }
 
@@ -989,18 +412,7 @@ export default class BackendAIData extends BackendAIPage {
     this.openDialog('add-folder-dialog');
   }
 
-  _folderExplorerDialog() {
-    this.openDialog('folder-explorer-dialog');
-  }
-
-  _mkdirDialog() {
-    this.shadowRoot.querySelector('#mkdir-name').value = '';
-    this.openDialog('mkdir-dialog');
-  }
-
   openDialog(id) {
-    //var body = document.querySelector('body');
-    //body.appendChild(this.$[id]);
     this.shadowRoot.querySelector('#' + id).show();
   }
 
@@ -1010,16 +422,6 @@ export default class BackendAIData extends BackendAIPage {
 
   _indexFrom1(index) {
     return index + 1;
-  }
-
-  _hasPermission(item, perm) {
-    if (item.permission.includes(perm)) {
-      return true;
-    }
-    if (item.permission.includes('w') && perm === 'r') {
-      return true;
-    }
-    return false;
   }
 
   _addFolder() {
@@ -1055,405 +457,14 @@ export default class BackendAIData extends BackendAIPage {
     this.closeDialog('add-folder-dialog');
   }
 
-  _getControlId(e) {
-    const controller = e.target;
-    const controls = controller.closest('#controls');
-    const folderId = controls.getAttribute('folder-id');
-    return folderId;
-  }
-
-  _infoFolder(e) {
-    const folderId = this._getControlId(e);
-    let job = window.backendaiclient.vfolder.info(folderId);
-    job.then((value) => {
-      this.folderInfo = value;
-      this.openDialog('info-folder-dialog');
-    }).catch(err => {
-      console.log(err);
-      if (err && err.message) {
-        this.notification.text = PainKiller.relieve(err.title);
-        this.notification.detail = err.message;
-        this.notification.show(true, err);
-      }
-    });
-  }
-
-  _deleteFolderDialog(e) {
-    this.deleteFolderId = this._getControlId(e);
-    this.shadowRoot.querySelector('#delete-folder-name').value = '';
-    this.openDialog('delete-folder-dialog');
-  }
-
-  _deleteFolderWithCheck() {
-    let typedDeleteFolderName = this.shadowRoot.querySelector('#delete-folder-name').value;
-    if (typedDeleteFolderName != this.deleteFolderId) {
-      this.notification.text = 'Folder name mismatched. Check your typing.';
-      this.notification.show();
-      return;
-    }
-    this.closeDialog('delete-folder-dialog');
-    this._deleteFolder(this.deleteFolderId);
-  }
-
-  _deleteFolder(folderId) {
-    let job = window.backendaiclient.vfolder.delete(folderId);
-    job.then((value) => {
-      this.notification.text = 'Folder is successfully deleted.';
-      this.notification.show();
-      this._refreshFolderList();
-    }).catch(err => {
-      console.log(err);
-      if (err && err.message) {
-        this.notification.text = PainKiller.relieve(err.title);
-        this.notification.detail = err.message;
-        this.notification.show(true, err);
-      }
-    });
-  }
-
-  /*Folder Explorer*/
-  _clearExplorer(path = this.explorer.breadcrumb.join('/'),
-                 id = this.explorer.id,
-                 dialog = false) {
-    let job = window.backendaiclient.vfolder.list_files(path, id);
-    job.then(value => {
-      this.shadowRoot.querySelector('#fileList-grid').selectedItems = [];
-      this.explorer.files = JSON.parse(value.files);
-      this.explorerFiles = this.explorer.files;
-      if (dialog) {
-        this.openDialog('folder-explorer-dialog');
-      }
-    });
-  }
-
-  _folderExplorer(e) {
-    let folderId = this._getControlId(e);
-    let explorer = {
-      id: folderId,
-      breadcrumb: ['.'],
-    };
-
-    this.explorer = explorer;
-    this._clearExplorer(explorer.breadcrumb.join('/'), explorer.id, true);
-  }
-
-  _enqueueFolder(e) {
-    const fn = e.target.getAttribute('name');
-    this.explorer.breadcrumb.push(fn);
-    this._clearExplorer();
-  }
-
-  _gotoFolder(e) {
-    const dest = e.target.getAttribute('dest');
-    let tempBreadcrumb = this.explorer.breadcrumb;
-    const index = tempBreadcrumb.indexOf(dest);
-
-    if (index === -1) {
-      return;
-    }
-
-    tempBreadcrumb = tempBreadcrumb.slice(0, index + 1);
-
-    this.explorer.breadcrumb = tempBreadcrumb;
-    this._clearExplorer(tempBreadcrumb.join('/'), this.explorer.id, false);
-  }
-
-  _mkdir(e) {
-    const newfolder = this.shadowRoot.querySelector('#mkdir-name').value;
-    const explorer = this.explorer;
-    let job = window.backendaiclient.vfolder.mkdir([...explorer.breadcrumb, newfolder].join('/'), explorer.id);
-    job.then(res => {
-      this.closeDialog('mkdir-dialog');
-      this._clearExplorer();
-    });
-  }
-
-  _isDir(file) {
-    return file.mode.startsWith("d");
-  }
-
-  /* File upload and download */
-  _addEventListenerDropZone() {
-    const dndZoneEl = this.shadowRoot.querySelector('#folder-explorer-dialog');
-    const dndZonePlaceholderEl = this.shadowRoot.querySelector('#dropzone');
-
-    dndZonePlaceholderEl.addEventListener('dragleave', () => {
-      dndZonePlaceholderEl.style.display = "none";
-    });
-
-    dndZoneEl.addEventListener('dragover', e => {
-      e.stopPropagation();
-      e.preventDefault();
-      e.dataTransfer.dropEffect = 'copy';
-      dndZonePlaceholderEl.style.display = "flex";
-      return false;
-    });
-
-    dndZoneEl.addEventListener('drop', e => {
-      e.stopPropagation();
-      e.preventDefault();
-      dndZonePlaceholderEl.style.display = "none";
-
-      let temp: any = [];
-      for (let i = 0; i < e.dataTransfer.files.length; i++) {
-        const file = e.dataTransfer.files[i];
-        if (file.size > 2 ** 20) {
-          console.log('File size limit (< 1 MiB)');
-        } else {
-          file.progress = 0;
-          file.caption = '';
-          file.error = false;
-          file.complete = false;
-          temp.push(file);
-          (this.uploadFiles as any).push(file);
-        }
-      }
-
-      for (let i = 0; i < temp.length; i++) {
-        this.fileUpload(temp[i]);
-        this._clearExplorer();
-      }
-    });
-  }
-
-  _uploadFileBtnClick(e) {
-    const elem = this.shadowRoot.querySelector('#fileInput');
-    if (elem && document.createEvent) {  // sanity check
-      const evt = document.createEvent("MouseEvents");
-      evt.initEvent("click", true, false);
-      elem.dispatchEvent(evt);
-    }
-  }
-
-  _uploadFileChange(e) {
-    const length = e.target.files.length;
-    for (let i = 0; i < length; i++) {
-      const file = e.target.files[i];
-
-      let text = "";
-      let possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-      for (let i = 0; i < 5; i++) text += possible.charAt(Math.floor(Math.random() * possible.length));
-
-      file.id = text;
-      file.progress = 0;
-      file.caption = '';
-      file.error = false;
-      file.complete = false;
-      (this.uploadFiles as any).push(file);
-    }
-
-    for (let i = 0; i < length; i++) {
-      this.fileUpload(this.uploadFiles[i]);
-    }
-
-    this.shadowRoot.querySelector('#fileInput').value = '';
-  }
-
-  fileUpload(fileObj) {
-    this._uploadFlag = true;
-    this.uploadFilesExist = this.uploadFiles.length > 0;
-    const path = this.explorer.breadcrumb.concat(fileObj.name).join("/");
-    let job = window.backendaiclient.vfolder.create_upload_session(path, fileObj, this.explorer.id);
-    job.then(url => {
-      const start_date = new Date().getTime();
-      const upload = new tus.Upload(fileObj, {
-        endpoint: url,
-        retryDelays: [0, 3000, 5000, 10000, 20000],
-        uploadUrl: url,
-        chunkSize: 15728640, // 15MB
-        metadata: {
-          filename: path,
-          filetype: fileObj.type
-        },
-        onError: function (error) {
-          console.log("Failed because: " + error)
-        },
-        onProgress: (bytesUploaded, bytesTotal) => {
-          if (!this._uploadFlag) {
-            upload.abort();
-            (this.uploadFiles as any)[(this.uploadFiles as any).indexOf(fileObj)].caption = `Canceling...`;
-            this.uploadFiles = this.uploadFiles.slice();
-            setTimeout(() => {
-              this.uploadFiles = [];
-              this.uploadFilesExist = false;
-            }, 1000);
-            return;
-          }
-
-          const now = new Date().getTime();
-          const speed: string = (bytesUploaded / (1024 * 1024) / ((now - start_date) / 1000)).toFixed(1) + "MB/s";
-          const estimated_seconds = Math.floor((bytesTotal - bytesUploaded) / (bytesUploaded / (now - start_date) * 1000));
-          let estimated_time_left = "Less than 10 seconds";
-          if (estimated_seconds >= 86400) {
-            estimated_time_left = "More than a day";
-          } else if (estimated_seconds > 10) {
-            const hour = Math.floor(estimated_seconds / 3600);
-            const min = Math.floor((estimated_seconds % 3600) / 60);
-            const sec = estimated_seconds % 60;
-            estimated_time_left = `${hour}:${min}:${sec}`;
-          }
-          const percentage = (bytesUploaded / bytesTotal * 100).toFixed(1);
-          (this.uploadFiles as any)[(this.uploadFiles as any).indexOf(fileObj)].progress = bytesUploaded / bytesTotal;
-          (this.uploadFiles as any)[(this.uploadFiles as any).indexOf(fileObj)].caption = `${percentage}% / Time left : ${estimated_time_left} / Speed : ${speed}`;
-          this.uploadFiles = this.uploadFiles.slice();
-        },
-        onSuccess: () => {
-          this._clearExplorer();
-          (this.uploadFiles as any)[(this.uploadFiles as any).indexOf(fileObj)].complete = true;
-          this.uploadFiles = this.uploadFiles.slice();
-          setTimeout(() => {
-            this.uploadFiles.splice((this.uploadFiles as any).indexOf(fileObj), 1);
-            this.uploadFilesExist = this.uploadFiles.length > 0 ? true : false;
-            this.uploadFiles = this.uploadFiles.slice();
-          }, 1000);
-        }
-      });
-      upload.start();
-    });
-  }
-
-  _cancelUpload(e) {
-    this._uploadFlag = false;
-  }
-
-  _downloadFile(e) {
-    let fn = e.target.getAttribute("filename");
-    let path = this.explorer.breadcrumb.concat(fn).join("/");
-    let job = window.backendaiclient.vfolder.request_download_token(path, this.explorer.id);
-    job.then(res => {
-      const token = res.token;
-      const url = window.backendaiclient.vfolder.get_download_url_with_token(token);
-      let a = document.createElement('a');
-      a.addEventListener('click', function (e) {
-        e.stopPropagation();
-      });
-      a.href = url;
-      a.download = fn;
-      document.body.appendChild(a); // we need to append the element to the dom -> otherwise it will not work in firefox
-      a.click();
-      a.remove();  //afterwards we remove the element again
-      URL.revokeObjectURL(url);
-    });
-  }
-
-  _openDeleteFileDialog(e) {
-    let fn = e.target.getAttribute("filename");
-    this.deleteFileDialog.filename = fn;
-    this.deleteFileDialog.files = [];
-    this.deleteFileDialog.show();
-  }
-
-  _openDeleteMultipleFileDialog(e?) {
-    this.deleteFileDialog.files = this.fileListGrid.selectedItems;
-    this.deleteFileDialog.filename = '';
-    this.deleteFileDialog.show();
-  }
-
-  _deleteFileWithCheck(e) {
-    let files = this.deleteFileDialog.files;
-    if (files.length > 0) {
-      let filenames: string[] = [];
-      files.forEach((file) => {
-        let filename = this.explorer.breadcrumb.concat(file.filename).join("/");
-        filenames.push(filename);
-      });
-      let job = window.backendaiclient.vfolder.delete_files(filenames, true, this.explorer.id);
-      job.then(res => {
-        this.notification.text = 'Files deleted.';
-        this.notification.show();
-        this._clearExplorer();
-        this.deleteFileDialog.hide();
-      });
-    } else {
-      if (this.deleteFileDialog.filename != '') {
-        let path = this.explorer.breadcrumb.concat(this.deleteFileDialog.filename).join("/");
-        let job = window.backendaiclient.vfolder.delete_files([path], true, this.explorer.id);
-        job.then(res => {
-          this.notification.text = 'File deleted.';
-          this.notification.show();
-          this._clearExplorer();
-          this.deleteFileDialog.hide();
-        });
-      }
-    }
-  }
-
-  _deleteFile(e) {
-    let fn = e.target.getAttribute("filename");
-    let path = this.explorer.breadcrumb.concat(fn).join("/");
-    let job = window.backendaiclient.vfolder.delete_files([path], true, this.explorer.id);
-    job.then(res => {
-      this.notification.text = 'File deleted.';
-      this.notification.show();
-      this._clearExplorer();
-    });
-  }
-
-  _humanReadableTime(d) {
-    const date = new Date(d * 1000);
-    const offset = date.getTimezoneOffset() / 60;
-    const hours = date.getHours();
-    date.setHours(hours - offset);
-    return date.toUTCString();
-  }
-
-  _isDownloadable(file) {
-    return true;
+  _refreshFolderList() {
+    // Send notification to folder objects
   }
 
   _hideDialog(e) {
     let hideButton = e.target;
     let dialog = hideButton.closest('wl-dialog');
     dialog.hide();
-  }
-
-  _shareFolderDialog(e) {
-    this.selectedFolder = this._getControlId(e);
-
-    this.openDialog('share-folder-dialog');
-  }
-
-  _modifyPermissionDialog(vfolder_id) {
-    window.backendaiclient.vfolder.list_invitees(vfolder_id)
-      .then(res => {
-        this.invitees = res.shared;
-        this.openDialog('modify-permission-dialog');
-      })
-  }
-
-  _shareFolder(e) {
-    // the .children property is an HtmlCollection. They don't have the map function like an array would
-    const emailHtmlCollection = this.shadowRoot.querySelector('#textfields').children;
-
-    // filter invalid and empty fields
-    const emailArray = Array.prototype.filter.call(emailHtmlCollection, e => !e.hasAttribute('invalid') && e.value !== '').map(e => e.value.trim());
-    const permission = 'r' + (this.shadowRoot.querySelector('#share-folder-write').checked ? 'w' : 'o');
-
-    if (emailArray.length === 0) {
-      this.notification.text = 'No valid emails were entered';
-      this.notification.show();
-      this.shadowRoot.querySelector('#share-folder-dialog').hide();
-      for (let element of emailHtmlCollection) {
-        element.value = '';
-      }
-      return;
-    }
-
-    window.backendaiclient.vfolder.invite(permission, emailArray, this.selectedFolder)
-      .then(res => {
-        let msg;
-        if (res.invited_ids && res.invited_ids.length > 0) {
-          msg = res.invited_ids.reduce((cur, val) => cur + val + " ", "") + (emailArray.length === 1 ? 'was' : 'were') + " successfully invited";
-        } else {
-          msg = "No one was invited";
-        }
-        this.notification.text = msg;
-        this.notification.show();
-        this.shadowRoot.querySelector('#share-folder-dialog').hide();
-        for (let element of emailHtmlCollection) {
-          element.value = '';
-        }
-      })
   }
 }
 declare global {
